Add mdoc:js modifier for Scala.js integration.
This commit adds a new modifier that makes it possible to write
interactive docs with Scala.js.

The following document gets compiled to JavaScript and is evaluated at
HTML load time instead of markdown generation time.
````md
```scala mdoc:js
println("Hello world!")
```
````

The modifier is published as a separate artifact `mdoc-js` because it
brings in a new dependency on the Scala.js compiler.